pingu: support for setting source-ip
This is needed so ip source routing works with shorewall

The problem is that when we change the default gw, removing a gateway that
went down, we are not able to ping remote host via that interface ...unless
we set the source ip and have an ip route table with source ip. Shorewall
sets that up.