pingu_host: bring interface down instantly if interface disapears
We need execute the down hook in case interface goes down if we have
gre tunnels bound to an interface.

If for example ppp0 disapears (pppd restart) and comes back before pingu
notices it, the gre interface will not be bound to anything. So we force
the "down" event instantly if we detect that the iface is gone.